Update Interim.md
Made some changes in Related Work section as per Myan Comments :
"Web 2.0 emphasis to e-democracy wherein multimodal, non-hierarchal, collaborative and deliberative networks can be built. 
I'm not quite sure what this means. This feels similar to some of the statements Peggy made about reducing the jargon in our report, so we either need to clearly define what we mean when we say each of these terms or cut this line of thought.

​[7:51] 
In the immediate section afterwards, who was conducting this study on web 2.0 applications? Was it an entity? An individual?"

Let me know if its ok.